+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2026-05-03
+
+### Added
+
+- Added `Connection.AutoNickRecoveryPostRegistration` to let applications disable automatic alternative `NICK` retries for 432/433/436/437 after registration.
+
+### Changed
+
+- Pre-registration nickname recovery remains enabled regardless of `AutoNickRecoveryPostRegistration`.
+
 ## [1.2.7] - 2026-05-02
 
 ### Added
@@ -125,6 +135,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+[1.3.0]: https://github.com/kofany/go-ircevo/compare/v1.2.7...v1.3.0
 [1.2.7]: https://github.com/kofany/go-ircevo/compare/v1.2.6...v1.2.7
 [1.2.6]: https://github.com/kofany/go-ircevo/compare/v1.2.5...v1.2.6
 [1.2.5]: https://github.com/kofany/go-ircevo/compare/v1.2.4...v1.2.5